rename : Kiosk_Lv5 -> Kiosk_Lv6 으로폴더 이동
